Move decorated commands into cogs
