+++ cogs/app_admin.py
@@ -1,0 +1,35 @@
+"""App-level administrative commands registered under the cogs namespace."""
+
+from __future__ import annotations
+
+from discord.ext import commands
+
+from c1c_coreops.helpers import tier
+from c1c_coreops.rbac import admin_only
+
+
+class AppAdmin(commands.Cog):
+    """Lightweight administrative utilities for bot operators."""
+
+    def __init__(self, bot: commands.Bot) -> None:
+        self.bot = bot
+
+    @tier("admin")
+    @commands.command(
+        name="ping",
+        hidden=True,
+        help="Quick admin check to confirm the bot is responsive.",
+    )
+    @admin_only()
+    async def ping(self, ctx: commands.Context) -> None:
+        """React with a paddle to confirm the bot processed the request."""
+
+        try:
+            await ctx.message.add_reaction("🏓")
+        except Exception:
+            # Reaction failures are non-fatal (missing perms, deleted message, etc.).
+            pass
+
+
+async def setup(bot: commands.Bot) -> None:
+    await bot.add_cog(AppAdmin(bot))
