+++ modules/recruitment/welcome.py
@@ -1,81 +1,7 @@
+"""Shim module preserving import compatibility for the recruitment welcome cog."""
+
 from __future__ import annotations
-from typing import Any, Dict, Optional
-from discord.ext import commands
 
-from modules.common import runtime as rt
-# NOTE: Do not import role ID constants from shared.config; not exported here.
-from c1c_coreops.helpers import tier
-from c1c_coreops.rbac import is_staff_member, is_admin_member
-from shared.sheets import async_facade as sheets
+from cogs.recruitment_welcome import WelcomeBridge, setup, staff_only
 
-# --- RBAC decorator (staff with fallback) -------------------------------------
-def staff_only():
-    """
-    Allow staff/admin via CoreOps roles. Also allow Discord 'Administrator'
-    permission as fallback (useful on fresh/dev guilds without CoreOps roles).
-    """
-    async def predicate(ctx: commands.Context) -> bool:
-        author = getattr(ctx, "author", None)
-        # CoreOps roles OR server Administrator (fallback)
-        if is_staff_member(author) or is_admin_member(author):
-            return True
-        if getattr(ctx, "_coreops_suppress_denials", False):
-            raise commands.CheckFailure("Staff only.")
-        try:
-            await ctx.reply("Staff only.")
-        except Exception:
-            pass
-        raise commands.CheckFailure("Staff only.")
-    return commands.check(predicate)
-
-class WelcomeBridge(commands.Cog):
-    """
-    Recruitment welcome command using cached templates.
-    Behavior matches legacy: it renders a template row and posts to the target channel.
-    """
-
-    def __init__(self, bot: commands.Bot):
-        self.bot = bot
-
-    @tier("staff")
-    @commands.command(name="welcome", usage="[clan] @mention")
-    @staff_only()
-    async def welcome(self, ctx: commands.Context, clan: Optional[str] = None, *, note: Optional[str] = None):
-        """
-        Post a templated welcome. Template rows are read from the cached 'templates' bucket.
-        Usage: !welcome [clan] @mention
-        """
-        templates = await sheets.get_cached_welcome_templates()
-        if not templates:
-            await ctx.send("⚠️ No welcome templates found. Try again after the next refresh.")
-            return
-
-        tag = (clan or "").strip().upper()
-        row: Optional[Dict[str, Any]] = None
-        for r in templates:
-            rtag = str(r.get("ClanTag", "")).strip().upper()
-            if rtag == tag:
-                row = r
-                break
-
-        if not row:
-            await ctx.send(f"⚠️ No template configured for clan tag `{tag}`.")
-            return
-
-        text = str(row.get("Message", "")).strip()
-        if not text:
-            await ctx.send(f"⚠️ Template for `{tag}` is missing a 'Message' field.")
-            return
-        if note:
-            text = f"{text}\n\n{note}"
-
-        await ctx.send(text)
-
-        # Unified log line
-        try:
-            await rt.send_log_message(f"[welcome] actor={ctx.author} clan={tag} channel=#{getattr(ctx.channel, 'name', '?')}")
-        except Exception:
-            pass
-
-async def setup(bot: commands.Bot):
-    await bot.add_cog(WelcomeBridge(bot))
+__all__ = ["WelcomeBridge", "setup", "staff_only"]
